requirements.txt
discord.py
requests
quart
hypercorn
